Tags: add HasAttributes method.
This methods will allow to check if current token has attributes.

As current attribute processing way is to parse them one-by-one lazily - this function can't tell if there are more unread attributes, only the fact that when this token was returned it had unread attributes. Maybe in the future this behavior will change.